+++ pyproject.toml
@@ -1,24 +1,21 @@
 [tool.poetry]
-name = "project-poetry"
+name = "project"
 version = "0.1.0"
 description = ""
-authors = ["GosuOoops <evgenii_neo@mail.ru>"]
+authors = ["Eremeev Sergey <saeremeev@gmail.com>"]
+readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.11"
-pyyaml = "^6.0"
-flask  = "^2.3"
-Flask-SQLAlchemy = "^3.0"
-psycopg2 = "^2.9"
+pyyaml = "6.*"
+flask = "*"
 
 [tool.poetry.group.dev]
 optional = true
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.3.0"
 black = "^23.3.0"
 
-
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
